Feat: 영상별 유입 경로, 시청자층, 사용기기 데이터 수집 스케쥴링 구현

+++ back_end/src/main/java/com/cm/astb/entity/DeviceAnalysisId.java
@@ -21,7 +21,7 @@ public class DeviceAnalysisId implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Column(name = "VIDEO_ID", nullable = false)
-    private Integer videoId;
+    private Long videoId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "DEVICE_TYPE", nullable = false)
